Improving download with multithread
This PR change the download function for the flickr and bing search
class. Was added a new way to multithreading the data URL's for a fast
download increasing a way better performance with big images searchs.

To test this change just run the idt init with all needed parameters
and run idt build at the end, download time will be
more faster than usual.